move out avatarstack to own file
